Unify worker build scripts for better error messages
